requirements.txt
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
seaborn==0.13.0
matplotlib==3.8.2
openpyxl==3.1.2
pyarrow==13.0.0
apache-airflow==2.7.3
apache-airflow-providers-postgres==6.6.1
pyspark==3.4.1
